# ex1/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Jogos de Tabuleiro",
    "version": "1.0.0",
    "description": "API desenvolvida para o exame de EngWeb2026."
  },
  "servers": [
    {
      "url": "http://localhost:17000"
    }
  ],
  "paths": {
    "/jogos": {
      "get": {
        "summary": "Obtém todos os jogos ou filtra por editora",
        "parameters": [
          {
            "in": "query",
            "name": "editora",
            "schema": { "type": "string" },
            "description": "Filtrar por nome da editora (ex: KOSMOS)"
          }
        ],
        "responses": { "200": { "description": "Sucesso" } }
      },
      "post": {
        "summary": "Adiciona um novo jogo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": { "201": { "description": "Criado com sucesso" } }
      }
    },
    "/jogos/{id}": {
      "get": {
        "summary": "Obtém um jogo pelo ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID do jogo (ex: catan)"
          }
        ],
        "responses": { "200": { "description": "Sucesso" }, "404": { "description": "Jogo não encontrado" } }
      },
      "put": {
        "summary": "Atualiza um jogo pelo ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": { "200": { "description": "Atualizado com sucesso" } }
      },
      "delete": {
        "summary": "Elimina um jogo pelo ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "Eliminado com sucesso" } }
      }
    },
    "/autores": {
      "get": {
        "summary": "Obtém a lista de autores",
        "responses": { "200": { "description": "Sucesso" } }
      }
    },
    "/categorias": {
      "get": {
        "summary": "Obtém a lista de categorias",
        "responses": { "200": { "description": "Sucesso" } }
      }
    }
  }
}